Refactor: Restructure files and update dependencies
-   Deleted `lib/core/api_result.dart`, `lib/presentation/screen/home_screen/tabs/cart_screen/cart_screen.dart`, `lib/presentation/screen/home_screen/tabs/category_screen/category_screen.dart`, `lib/presentation/screen/home_screen/tabs/home_screen/home_screen.dart`, `lib/presentation/screen/home_screen/layout_view.dart`, `lib/presentation/screen/home_screen/tabs/profile_screen/profile_screen.dart`
-   Created `lib/presentation/home/tabs/cart/cart_screen.dart`, `lib/presentation/home/tabs/category/category_screen.dart`, `lib/presentation/home/tabs/home/home_screen.dart`, `lib/presentation/home/layout_view.dart`, `lib/presentation/home/tabs/profile/profile_screen.dart`
-   Renamed `lib/api/api_excuter.dart` to `lib/data/api_excuter.dart`, `lib/api/web_services/WebServices.dart` to `lib/data/web_services/WebServices.dart`
- Renamed `lib/presentation/screen/auth_screen/login/login_screen.dart` to `lib/presentation/auth/screens/login_screen.dart`
- Renamed `lib/domain/common/result.dart` to `lib/domain/common/api_result.dart`
- Updated imports and references to reflect file renames and deletions.
-  Updated StringManager, added baseUrl in WebServices,  refactor all files to new structure.
- Added new Cubits and states files.
