+++ lib/core/utils/resources/string_manager.dart
@@ -21,7 +21,7 @@ static const String hintEnterEmailToResetPassword = 'please Enter your email ass
   static const String phoneNumberError = 'Phone number must be 11 digits long and start with 01.';
   static const String rePasswordError = 'Passwords do not match';
   static const String login = 'Login';
-  static const String continueasguest = 'Continue as guest';
+  static const String continueAsGuest = 'Continue as guest';
   static const String signup = 'Sign up';
   static const String home = 'Home';
   static const String profile = 'Profile';
