+++ lib/data/api_excuter.dart
@@ -1,12 +1,12 @@
 
 import 'package:dio/dio.dart';
-import 'package:flowery/domain/common/result.dart';
+import 'package:flowery/domain/common/api_result.dart';
 import 'package:flowery/domain/entity/error_model.dart';
 import 'package:flowery/domain/exception/server_error.dart';
 
 typedef ApiCall<T> = Future<T> Function();
 
-Future<Result<T>> executeApi<T>(ApiCall<T> apiCall) async {
+Future<ApiResult<T>> executeApi<T>(ApiCall<T> apiCall) async {
   try {
     var result = await apiCall.call();
     return Success(result);
@@ -18,26 +18,26 @@ Future<Result<T>> executeApi<T>(ApiCall<T> apiCall) async {
       case DioExceptionType.receiveTimeout:
       case DioExceptionType.connectionTimeout:
         {
-          return Errorr(NetworkError());
+          return Error(NetworkError());
         }
       case DioExceptionType.badResponse:
         {
           var responseCode = ex.response?.statusCode ?? 0;
           var errorModel = ErrorModel.fromJson(ex.response?.data);
           if (responseCode >= 400 && responseCode < 500) {
-            return Errorr(ClientError(errorModel));
+            return Error(ClientError(errorModel));
           }
           if (responseCode >= 500 && responseCode < 600) {
-            return Errorr(ServerError(errorModel));
+            return Error(ServerError(errorModel));
           }
-          return Errorr(Exception("Something went wrong!"));
+          return Error(Exception("Something went wrong!"));
         }
       default:
         {
-          return Errorr(Exception("Something went wrong!"));
+          return Error(Exception("Something went wrong!"));
         }
     }
   } on  Exception catch(ex){
-    return Errorr(ex);
+    return Error(ex);
   }
 }
