+++ lib/data/web_services/WebServices.dart
@@ -1,14 +1,13 @@
 import 'package:dio/dio.dart';
-import 'package:injectable/injectable.dart';
 import 'package:retrofit/error_logger.dart';
 import 'package:retrofit/http.dart';
 
 part 'WebServices.g.dart';
 
-@RestApi()
+@RestApi(baseUrl: 'https://flower.elevateegy.com/')
 abstract class WebServices{
 
-  factory WebServices(Dio dio, {String baseUrl}) = _WebServices;
+  factory WebServices(Dio dio,) = _WebServices;
 
 
 }
